Fix disconnection detection in liburbi.
	* uclient.cc: Block SIGPIPE, include the socket FD in the error
	 FD_SET in select, treat a return value of 0 from recv as an error
	 (nominal disconnection).
	* uconversion.cc: Replace an int by a long to fix a warning on amd64.

git-svn-id: file:///srv/svn/liburbi-cpp/trunk@516 b8db33d8-8104-0410-b827-8403f563b8eb

+++ sdk-remote/src/liburbi/uconversion.cc
@@ -535,7 +535,7 @@ namespace urbi
   void copy(S* src, D* dst,
 	    int sc, int dc, int sr, int dr, int count, bool sf, bool df)
   {
-    int shift = 8 * (sizeof (S) - sizeof (D));
+    long shift = 8 * (sizeof (S) - sizeof (D));
     for (int i = 0; i < count; ++i)
     {
       float soffset = (float)i * ((float)sr / (float)dr);
